renovate: Update GitHub Actions via semver tags *with* SHAs
There are two primary ways to have Renovate keep thirdparty GitHub
Actions up-to-date:

1. pin to a semantic version (`uses: foo/bar@v1.2.3`), or
2. pin to a commit hash (`uses: foo/bar@abcdef012345`)

Approach (1) is much more understandable at-a-glance and more compatible
with Renovate's "show the changelog" feature.  Also, it avoids depending
directly on the bleeding edge of the `master` branch of these actions.

On the other hand, (2) is much better for security and reproducibility,
since repo authors are free to overwrite tags whenever they wish.

I noticed that https://github.com/astral-sh/uv was using a hybrid
approach where they were using a syntax like

3. pin to both (`uses: foo/bar@abcdef012345 # v1.2.3`)

which seems to be the best of both worlds.  So this patch is just
copypasta from
https://github.com/astral-sh/uv/blob/574aa1ef110ef08293512eb200bd6881bb738179/.github/renovate.json5#L25-L35